Rename Sort.js file
